+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * **Larger change:** Align session expiration logic with [connect-redis](https://github.com/tj/connect-redis/blob/30efd159103ace270c844a5967428b43e7b8ba4a/migration-to-v4.md#changes-to-ttl-management) which in turned aligned with connect-mongo. Fixes #54
 * **Feature:** New option: `disableTouch`. Disables updating of TTL in database on touch. Fixes #55.
+* **Feature:** New option: `createTableIfMissing`. When set, the session table will be automatically created if missing. Fixes #158 and #173. Thanks @aadeshmisra!
 
 ## 6.2.1 (2020-08-19)